refactor: dpvi moved into separate directory
